+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Features
+^^^^^^^^
+* Added a new option `--django-debug` to set the DEBUG setting to True prior to
+  running tests.
+
 3.1.2
 -----